--- build.gradle
@@ -1,49 +1,0 @@
-plugins {
-	id 'org.springframework.boot' version '2.7.3'
-	id 'io.spring.dependency-management' version '1.0.13.RELEASE'
-	id 'java'
-}
-
-group = 'com.crg.todo'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-// spring boot
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'org.springframework.boot:spring-boot-starter-security'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-
-// spring boot test
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.springframework.security:spring-security-test'
-	testCompileOnly 'org.projectlombok:lombok:1.18.24'
-	testAnnotationProcessor 'org.projectlombok:lombok:1.18.24'
-
-// database
-	//	runtimeOnly 'org.postgresql:postgresql'
-	runtimeOnly 'org.mariadb.jdbc:mariadb-java-client'
-
-// other
-	implementation 'org.json:json:20141113'
-	compileOnly 'org.projectlombok:lombok:1.18.24'
-	annotationProcessor 'org.projectlombok:lombok:1.18.24'
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
-
-apply plugin: 'maven-publish'
-
-publishing {
-	publications {
-		customLibrary(MavenPublication) {
-			from components.java
-		}
-	}
-}
