+++ docker-compose.yml
@@ -1,30 +1,88 @@
-# Use root/example as user/password credentials
-version: '3.1'
+version: '3.9'
+
+# The previous config
+# services:
+#   # api_service:
+#   #   build: .
+#   #   restart: always
+#   #   ports:
+#   #     - 8080:8080
+#   #   depends_on:
+#   #     - maria_db
+#   #   links:
+#   #     - maria_db:app_db
+
+#   maria_db:
+#     image: mariadb
+#     restart: always
+#     ports:
+#       - 3306:3306
+#     environment:
+#       MYSQL_DATABASE: todolist
+#       MYSQL_USER: root
+#       MYSQL_PASSWORD: example
+#       MYSQL_ROOT_PASSWORD: example
+
+#   adminer:
+#     image: adminer
+#     restart: always
+#     ports:
+#       - 9000:8080
 
 services:
-  # api_service:
-  #   build: .
-  #   restart: always
+  # backend:
+  #   build: ./server
+  #   volumes:
+  #     - ./server:/server
   #   ports:
-  #     - 8080:8080
+  #     - "8000:8000"
+  #   environment:
+  #     - SECRET_KEY=supersecret
+  #     - ENVIRONMENT=dev
+  #     - DB_HOST=db
+  #     - DB_NAME=postgres
+  #     - DB_USER=postgres
+  #     - DB_PASSWORD=password
   #   depends_on:
-  #     - maria_db
-  #   links:
-  #     - maria_db:app_db
+  #     - db
 
-  maria_db:
-    image: mariadb
-    restart: always
+  # frontend:
+  #   build: ./client
+  #   volumes:
+  #     - ./client:/client
+  #     # One-way volume to use node_modules from inside image
+  #     - /client/node_modules
+  #   ports:
+  #     - "3000:3000"
+  #   environment:
+  #     - NODE_ENV=development
+  #   depends_on:
+  #     - backend
+  #   command: npm start
+
+# Use root/example as user/password credentials
+  db:
+    image: postgres:12-alpine
     ports:
-      - 3306:3306
+      - "5432:5432"
+    restart: always
     environment:
-      MYSQL_DATABASE: todolist
-      MYSQL_USER: root
-      MYSQL_PASSWORD: example
-      MYSQL_ROOT_PASSWORD: example
+      - POSTGRES_DB=postgres
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=password
 
   adminer:
     image: adminer
     restart: always
+    depends_on:
+      - db
     ports:
-      - 9000:8080
+      - "9000:8080"
+
+  mailhog:
+    image: mailhog/mailhog
+    logging:
+      driver: "none" # disable saving logs
+    ports:
+      - 1025:1025 # smtp server
+      - 8025:8025 # web ui
