+++ server/src/main/java/nextmainfocus/tasks_list/Sorting.java
@@ -1,0 +1,5 @@
+package nextmainfocus.tasks_list;
+
+public enum Sorting {
+    ASC, DESC
+}
