+++ server/src/main/resources/application-prod.yml
@@ -1,0 +1,6 @@
+logging:
+    level:
+        org:
+            springframework: info
+        jlogicsoftware:
+            nextmainfocus: debug
