+++ server/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+spring:
+    profiles:
+        active: dev
