+++ server/src/test/java/nextmainfocus/ApplicationTests.java
@@ -1,0 +1,24 @@
+// package nextmainfocus;
+
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.boot.test.context.SpringBootTest;
+
+// import nextmainfocus.tasks_list.TasksController;
+
+// @SpringBootTest class ApplicationTests {
+
+//     @Autowired TasksController tasksController;
+
+//     // @Test void getTasksList() {
+// //        ArrayList<Task>
+// //                body = (ArrayList<Task>) tasksController
+// //                        .findTasks("", "", "")
+// //                        .getBody();
+// //        Integer status = tasksController
+// //                            .findTasks("", "", "")
+// //                            .getStatusCodeValue();
+// //
+// //        assertThat(tasksController).isNotNull();
+// //        assertThat(status).isEqualTo(200);
+//     // }
+// }
