--- src/main/java/com/crg/todo/security/entity/Account.java
@@ -1,45 +1,0 @@
-package com.crg.todo.security.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Account {
-    @Getter
-    @Setter
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @Getter
-    @Setter
-    private String username;
-    @Getter
-    @Setter
-    private String password;
-    @Getter
-    @Setter
-    private String role;
-
-// methods below are from Spring Security and not used for now
-
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    public boolean isEnabled() {
-        return true;
-    }
-}
