--- src/main/java/com/crg/todo/tasks_list/Sorting.java
@@ -1,5 +1,0 @@
-package com.crg.todo.tasks_list;
-
-public enum Sorting {
-    ASC, DESC
-}
