--- src/main/resources/application-local.yml
@@ -1,9 +1,0 @@
-##we're not using this config
-#
-#client:
-#  url: '*'
-#spring:
-#  datasource:
-#    url: jdbc:mariadb://localhost:3306/todolist
-#    username: root
-#    password: example
