--- src/main/resources/application.yml
@@ -1,17 +1,0 @@
-##we're not using this config
-#
-#client:
-#    url: '*'
-#spring:
-#    datasource:
-#        driver-class-name: org.mariadb.jdbc.Driver
-#        url: jdbc:mariadb://localhost:3306/todolist
-#        username: root
-#        password: example
-#    jpa:
-#        hibernate:
-#            ddl-auto: none
-#        properties:
-#            hibernate:
-#                dialect: org.hibernate.dialect.MySQL5InnoDBDialect
-#        show-sql: true
